+++ src/app/(auth)/master-data/uraian-tugas/[id]/page.tsx
@@ -69,6 +69,15 @@ export default async function EditUraianTugasPage({ params }: { params: Promise<
             tugas_pokok_id: uraianTugas.tugas_pokok_id,
             detil_tugas_id: uraianTugas.detil_tugas_id ?? "",
             jabatan_id: uraianTugas.jabatan_id ?? "",
+            std_sumber_bukti: uraianTugas.std_sumber_bukti ?? undefined,
+            std_kondisi: uraianTugas.std_kondisi ?? undefined,
+            std_frekuensi_teks: uraianTugas.std_frekuensi_teks ?? undefined,
+            std_durasi_per_kali: uraianTugas.std_durasi_per_kali ?? undefined,
+            std_jam_per_minggu: uraianTugas.std_jam_per_minggu ?? undefined,
+            std_peak4w_hours: uraianTugas.std_peak4w_hours ?? undefined,
+            std_ai_mode: uraianTugas.std_ai_mode ?? undefined,
+            std_va_type: uraianTugas.std_va_type ?? undefined,
+            std_dcs_flag: uraianTugas.std_dcs_flag ?? undefined,
           }}
           editId={id}
         />